+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added test support for `azure` (#146)
+- Added support for `delta` tables on S3 (#24)
 
 
 ## [0.10.1] - 2024-04-19