+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+
+## [0.10.2] - 2024-05-05
+
+### Added
+
 - Added import glue (#166)
 - Added test support for `azure` (#146)
 - Added support for `delta` tables on S3 (#24)